升级memory profiler到0.25.1

+++ Unity3d/Assets/Scripts/Editor/ScriptableDsl/ResourceProcess/MemoryProfiler/EditorCoroutine/EditorWaitForSeconds.cs
@@ -29,7 +29,7 @@
         /// <summary>
         /// The time to wait in seconds.
         /// </summary>
-        public double WaitTime { get; }
+        public float WaitTime { get; }
 
         /// <summary>
         /// Creates a instruction object for yielding inside a generator function.
